+++ thetawave-mobs/src/attributes/mod.rs
@@ -1,9 +1,9 @@
 mod data;
 mod plugin;
 
+pub use data::MobType;
 pub(crate) use data::{
     JointedMob, JointsComponent, MobAttributesComponent, MobAttributesResource, MobDecorationType,
     MobSpawnerComponent, ProjectileSpawnerComponent,
 };
-pub use data::{MobType, SpawnMobEvent};
 pub(crate) use plugin::ThetawaveAttributesPlugin;
